% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhm_getData.R
\name{nhm_getData}
\alias{nhm_getData}
\title{Search site-level information from the PREDICTS database}
\usage{
nhm_getData(what = "refs")
}
\arguments{
\item{what}{which file to download (`refs`, `bibtex`, `summary`, `data`). 
See details for more information.}
}
\value{
list of 5 entries (resource_id, fields, records, _links, and total)
}
\description{
Search site-level information from the PREDICTS database
}
\details{
This function provides a method to download information from the 
 PREDICTS 2016 database. The files included in the data store are a data.frame
 of references accessible by setting the argument what to `refs`, a bibtex
 file that can be downloaded by setting what to `bibtex`, a summary of the 
 survey sites accesible by setting what to `summary`, and the full dataset,
 accessible by setting what to `data`.
}
