% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhm_references.R
\name{nhm_references}
\alias{nhm_references}
\title{Access or download (bib) PREDICT database references}
\usage{
nhm_references(bib = FALSE, limit = NULL)
}
\arguments{
\item{bib}{download bibtex file containing all the references (boolean)}

\item{limit}{number of records to download (if NULL, set to max value = 5000)}
}
\value{
outputs a data.frame of 18 columns containing 
information on literature cited as part of the PREDICT database
}
\description{
Access or download (bib) PREDICT database references
}

