% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhm_species.R
\name{nhm_species}
\alias{nhm_species}
\title{Access}
\usage{
nhm_species(limit = NULL)
}
\arguments{
\item{limit}{number of records to download (if NULL, set to max value = 5000)}
}
\value{
output
}
\description{
Access
}

