% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhm_summary.R
\name{nhm_summary}
\alias{nhm_summary}
\title{Get information on entire data package}
\usage{
nhm_summary(id = NULL)
}
\arguments{
\item{id}{dataset id. Can be used to search other data packages available in the NHM if 
you'd like (default = NULL, which corresponds to the PREDICTS database id).}
}
\value{
data.frame contianing information on all files in the PREDICTS database repository.
}
\description{
Get information on entire data package
}

