% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHMpredict-package.R
\docType{package}
\name{NHMpredict-package}
\alias{NHMpredict-package}
\alias{NHMpredict}
\title{Access the 2016 PREDICTS database}
\description{
Access the 2016 PREDICTS database

The package began as an attempt to programmatically access the London Natural 
History Museum's Data Portal, and specifically the 2016 PREDICTS database. For now, 
it allows a way to access only the PREDICTS database, though functionality may be
added in the future to search and access other Data Portal sets.
}
\author{
Tad Dallas \email{tad.a.dallas@gmail.com}
}
