% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhm_site.R
\name{nhm_site}
\alias{nhm_site}
\title{Access stuff}
\usage{
nhm_site(limit = NULL)
}
\arguments{
\item{limit}{number of records to download (if NULL, set to max value = 5000)}
}
\value{
output
}
\description{
Access stuff
}

